+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added modifiable arena state names to language
 * Added configurable map selector items
 * Added legacy support
+* Cached armor stands will be removed after block break if type is matches
 
 ### 1.7.4 Release (25.01.2021 - 28.01.2021)
 * Fix stack trace array length is 0 when trying to send exception to service